backend/requirements.txt
Flask
flask-cors
python-dotenv
pymongo
Flask-PyMongo
mongoengine
PyJWT
Werkzeug
